Move YouTube component form inline instead of using a tooltip
